src/requirements.txt
beautifulsoup4==4.7.1
bleach==3.1.0
Django==2.2.2
django-import-export==1.2.0
django-mysql==3.2.0
djangorestframework==3.9.4
lxml==4.3.4
pandas==0.24.2
PyMySQL==0.9.3
python-dateutil==2.8.0
requests==2.22.0
Scrapy==1.6.0
scrapy-djangoitem==1.1.1
scrapy-proxy-pool==0.1.5
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
selenium==3.141.0
urllib3==1.25.3
user-agents==2.0
w3lib==1.20.0
django-filter==2.2.0
django-background-tasks==1.2.0
django-mathfilters==0.4.0
django-widget-tweaks==1.4.5
django-simple-history==2.7.3
geopandas==0.6.1
matplotlib==3.1.1
celery==4.3.0
django-celery-beat==1.5.0
django-celery-monitor==1.1.2
django-celery-results==1.1.2
